requirements.txt
# Gemini API
google-generativeai>=0.8.0

# PDF parsing
pymupdf>=1.24.0         # fitz - fast, accurate PDF text extraction
pdfplumber>=0.11.0      # good for tables in PDFs

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Data processing
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# NLP / text processing
tiktoken>=0.7.0

# Utilities
python-multipart>=0.0.9  # file upload support for FastAPI
